# shifter/shifter_platform/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shifter Platform API",
    "version": "v1",
    "description": "Authenticated HTTP/JSON API for Shifter platform integrations."
  },
  "paths": {
...
    "/api/v1/mission-control/agents/": {
      "get": {
        "operationId": "api_v1_mission_control_agents_list",
        "description": "Return agents available to the authenticated actor.",
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:read"
        ]
      }
    },
...
    "/api/v1/mission-control/range/": {
      "get": {
        "operationId": "api_v1_mission_control_range_retrieve",
        "description": "Return the active range and connection URLs for the request user.",
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentRangeResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:read"
        ]
      }
    },
...
    "/api/v1/mission-control/range/cancel/": {
      "post": {
        "operationId": "api_v1_mission_control_range_cancel",
        "description": "Run the configured range lifecycle service method.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:write"
        ]
      }
    },
    "/api/v1/mission-control/range/destroy/": {
      "post": {
        "operationId": "api_v1_mission_control_range_destroy",
        "description": "Run the configured range lifecycle service method.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:write"
        ]
      }
    },
    "/api/v1/mission-control/range/launch/": {
      "post": {
        "operationId": "api_v1_mission_control_range_launch",
        "description": "Validate input and create a range for the authenticated actor.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRange"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRange"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRange"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchRangeResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request was throttled."
          }
        },
        "x-required-scopes": [
          "mission_control:range:write"
        ]
      }
    },
    "/api/v1/mission-control/range/pause/": {
      "post": {
        "operationId": "api_v1_mission_control_range_pause",
        "description": "Run the configured range lifecycle service method.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:write"
        ]
      }
    },
    "/api/v1/mission-control/range/resume/": {
      "post": {
        "operationId": "api_v1_mission_control_range_resume",
        "description": "Run the configured range lifecycle service method.",
        "tags": [
          "mission-control"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/RangeLifecycle"
              }
            }
          }
        },
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:write"
        ]
      }
    },
    "/api/v1/mission-control/ranges/": {
      "get": {
        "operationId": "api_v1_mission_control_ranges_list",
        "description": "Return the authenticated actor's Mission Control range history, newest first.",
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RangeHistoryResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:read"
        ]
      }
    },
    "/api/v1/mission-control/scenarios/": {
      "get": {
        "operationId": "api_v1_mission_control_scenarios_list",
        "description": "Return scenarios available to the authenticated actor.",
        "tags": [
          "mission-control"
        ],
        "security": [
          {
            "ApiTokenAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication failed."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Permission denied."
          }
        },
        "x-required-scopes": [
          "mission_control:range:read"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentListResponse": {
        "type": "object",
        "description": "Response body for ``AgentListView.get``.",
        "properties": {
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentListItem"
            }
          }
        },
        "required": [
          "agents"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "description": "Schema-only component for the canonical platform API error envelope.\n\nMirrors :func:`shared.api.errors.api_exception_handler` /\n:func:`shared.api.errors.api_error_response` output. It describes the wire\ncontract; it never runs at request time.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CurrentRangeResponse": {
        "type": "object",
        "description": "Response body for ``CurrentRangeView.get``.",
        "properties": {
          "has_range": {
            "type": "boolean"
          },
          "range": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RangePresentation"
              }
            ],
            "nullable": true
          },
          "connection_urls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectionUrl"
            }
          },
          "aces_projection": {
            "type": "object",
            "additionalProperties": {},
            "nullable": true
          },
          "aces_participant_runtime": {
            "type": "object",
            "additionalProperties": {},
            "nullable": true
          }
        },
        "required": [
          "aces_participant_runtime",
          "aces_projection",
          "connection_urls",
          "has_range",
          "range"
        ]
      },
...
      "RangeLifecycle": {
        "type": "object",
        "description": "Validate range cancel/destroy/pause/resume requests.",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "range_id": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "description": "Generic ``{\"success\": true}`` response body.\n\nShared by the range lifecycle mutations (cancel/destroy/pause/resume),\nupload cancel, and credential delete — every Mission Control mutation\nendpoint that has no payload beyond a boolean acknowledgement.",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
    },
    "securitySchemes": {
      "ApiTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "shf",
        "description": "Platform API token with scopes from shared.api_tokens.scopes."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  }
}